my_dag_project/requirements/beam_requirements.txt

apache-beam[gcp]
google-cloud-bigtable
pyarrow
